+++ src/DDLQueryBuilder.php
@@ -8,11 +8,12 @@ use Exception;
 use Throwable;
 use Yiisoft\Db\Exception\InvalidArgumentException;
 use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\QueryBuilder\AbstractDDLQueryBuilder;
-use Yiisoft\Db\Schema\Builder\ColumnInterface;
+use Yiisoft\Db\Schema\Column\ColumnInterface;
 
 use function array_diff;
+use function implode;
+use function preg_replace;
 
 /**
  * Implements a (Data Definition Language) SQL statements for MSSQL Server.
@@ -53,43 +54,39 @@ final class DDLQueryBuilder extends AbstractDDLQueryBuilder
      */
     public function alterColumn(string $table, string $column, ColumnInterface|string $type): string
     {
-        $sqlAfter = [];
         $columnName = $this->quoter->quoteColumnName($column);
         $tableName = $this->quoter->quoteTableName($table);
-        $constraintBase = preg_replace('/[^a-z0-9_]/i', '', $table . '_' . $column);
+        $constraintBase = preg_replace('/\W/', '', $table . '_' . $column);
 
-        if ($type instanceof ColumnInterface) {
-            $type->setFormat('{type}{length}{notnull}{append}');
-
-            /** @psalm-var mixed $defaultValue */
-            $defaultValue = $type->getDefault();
-            if ($defaultValue !== null || $type->isNotNull() === false) {
-                $sqlAfter[] = $this->addDefaultValue(
-                    $table,
-                    "DF_{$constraintBase}",
-                    $column,
-                    $defaultValue instanceof Expression ? $defaultValue : new Expression((string)$defaultValue)
-                );
-            }
-
-            $checkValue = $type->getCheck();
-            if ($checkValue !== null) {
-                $sqlAfter[] = "ALTER TABLE {$tableName} ADD CONSTRAINT " .
-                    $this->quoter->quoteColumnName("CK_{$constraintBase}") .
-                    ' CHECK (' . ($defaultValue instanceof Expression ? $checkValue : new Expression($checkValue)) . ')';
-            }
-
-            if ($type->isUnique()) {
-                $sqlAfter[] = "ALTER TABLE {$tableName} ADD CONSTRAINT " . $this->quoter->quoteColumnName("UQ_{$constraintBase}") . " UNIQUE ({$columnName})";
-            }
+        if (is_string($type)) {
+            $type = $this->schema->getColumnFactory()->fromDefinition($type);
         }
 
-        /** @psalm-suppress DeprecatedMethod */
-        return implode("\n", [
+        $columnDefinitionBuilder = $this->queryBuilder->getColumnDefinitionBuilder();
+        $statements = [
             $this->dropConstraintsForColumn($table, $column, 'D'),
-            "ALTER TABLE $tableName ALTER COLUMN $columnName {$this->queryBuilder->getColumnType($type)}",
-            ...$sqlAfter,
-        ]);
+            "ALTER TABLE $tableName ALTER COLUMN $columnName " . $columnDefinitionBuilder->buildAlter($type),
+        ];
+
+        if ($type->hasDefaultValue()) {
+            $defaultValue = $type->dbTypecast($type->getDefaultValue());
+            $statements[] = $this->addDefaultValue($table, "DF_$constraintBase", $column, $defaultValue);
+        }
+
+        $checkValue = $type->getCheck();
+        if (!empty($checkValue)) {
+            $statements[] = "ALTER TABLE $tableName ADD CONSTRAINT "
+                . $this->quoter->quoteColumnName("CK_$constraintBase")
+                . " CHECK ($checkValue)";
+        }
+
+        if ($type->isUnique()) {
+            $statements[] = "ALTER TABLE $tableName ADD CONSTRAINT "
+                . $this->quoter->quoteColumnName("UQ_$constraintBase")
+                . " UNIQUE ($columnName)";
+        }
+
+        return implode("\n", $statements);
     }
 
     /**
@@ -117,6 +114,20 @@ final class DDLQueryBuilder extends AbstractDDLQueryBuilder
         }
 
         return $command;
+    }
+
+    public function createIndex(
+        string $table,
+        string $name,
+        array|string $columns,
+        string|null $indexType = null,
+        string|null $indexMethod = null
+    ): string {
+        return 'CREATE ' . (!empty($indexType) ? $indexType . ' ' : '') . 'INDEX '
+            . $this->quoter->quoteTableName($name) . ' ON '
+            . $this->quoter->quoteTableName($table)
+            . (!empty($columns) ? ' (' . $this->queryBuilder->buildColumns($columns) . ')' : '')
+            . (!empty($indexMethod) ? " USING $indexMethod" : '');
     }
 
     /**